+++ PostHog/PostHogConfig.swift
@@ -120,6 +120,17 @@ public typealias BeforeSendBlock = (PostHogEvent) -> PostHogEvent?
         set { setSurveys(newValue) }
     }
 
+    @available(iOS 15.0, *)
+    @available(watchOS, unavailable, message: "Surveys are only available on iOS 15+")
+    @available(macOS, unavailable, message: "Surveys are only available on iOS 15+")
+    @available(tvOS, unavailable, message: "Surveys are only available on iOS 15+")
+    @available(visionOS, unavailable, message: "Surveys are only available on iOS 15+")
+    @_spi(Experimental)
+    @objc public var surveysConfig: PostHogSurveysConfig {
+        get { _surveysConfig }
+        set { setSurveysConfig(newValue) }
+    }
+
     // only internal
     var disableReachabilityForTesting: Bool = false
     var disableQueueTimerForTesting: Bool = false
@@ -181,6 +192,15 @@ public typealias BeforeSendBlock = (PostHogEvent) -> PostHogEvent?
         // Unlike swift, which enforces stricter safety rules, objc just displays a warning
         if #available(iOS 15.0, *) {
             _surveys = value
+        }
+    }
+
+    var _surveysConfig: PostHogSurveysConfig = .init() // swiftlint:disable:this identifier_name
+    private func setSurveysConfig(_ value: PostHogSurveysConfig) {
+        // protection against objc API availability warning instead of error
+        // Unlike swift, which enforces stricter safety rules, objc just displays a warning
+        if #available(iOS 15.0, *) {
+            _surveysConfig = value
         }
     }